notif: Sort user IDs correctly in pm_users.
This fell into a classic JS pitfall: when sorting an array of numbers,
the default comparison function is by string, and you need to pass an
explicit comparison function to sort the actual numbers.

This means that in b022e9c1d we actually broke this invariant which
was probably previously always true in practice, oops.  Fortunately we
generally avoid actually relying on that invariant; see comment on
pmNarrowFromEmails (previously groupNarrow) in src/utils/narrow.js.