recipient: Fix some places that wrongly ignored bots and deactivated users.
Specifically cross-realm bots, like Welcome Bot and Notification Bot.

Such users can perfectly well appear in a link to a PM thread, and we
should handle those links correctly.

They can also appear in a PM thread given by a notification.  In the
case of deactivated users this isn't super likely, because IIUC the
message won't have been sent if the user was deactivated; but it could
happen if e.g. the user was deactivated after the message and before
our user got around to opening the notification.