user [nfc]: Use the full dataset on users in UserPickerCard.
This should have no effect on behavior, because this part of the UI
never shows in the first place the users that were left out of the
smaller map.  But, first, it's simpler to understand the code and be
confident it's right if we don't have to worry about checking that;
and second, having this special extra smaller map means we have to
wastefully construct a second map, duplicating references to all the
users that are in it.  So switch to looking up users in the normal,
complete map that contains all the users we know about.

Because the broader `allUsersByEmail` map contains values of the
slightly less specific type `UserOrBot` rather than `User`, this
also requires propagating that type through a range of other code.