user [nfc]: Cut out getUsersByEmail!  And unexport getUsersById.
Both of these data structures cause wasteful duplication of the
data we have on users; all the same user objects can be found in
`getAllUsersById`.  One of them moreover has been part of our use
of emails to identify users, which we're converting to user IDs.

We've just eliminated the last caller of `getUsersByEmail` outside
this selectors file itself; so we can convert the remaining users
to `getUsersById`, and then delete the selector entirely.

We also not long ago (in 9ba8e4476) removed the last caller of
`getUsersById` outside this file.  Converting the remaining uses
within the file isn't as trivial, so we'll leave that for a future
commit; but we can at least unexport it, except for tests, to avoid
adding further uses.