narrow [nfc]: Add getters for stream name and topic; use them.
As the jsdoc added here says, all these call sites should probably be
doing something else: where the code is really about a stream, or a
stream conversation, it should be getting passed down a stream name,
or stream name and topic (better yet, a stream *ID* and possibly a
topic) in the first place, rather than a Narrow.  Where on the other
hand the code naturally takes a whole Narrow because it handles other
kinds of narrows too, it should generally be using caseNarrow.

Fixing that would be its own set of refactors in a mostly
independent direction from this one, though, so for now we leave
that for a future project (or really series of small projects.)