narrow [nfc]: Switch the Narrow type to a reasonable structure!
Because of all the work we've done over the last few series to
encapsulate the details of the Narrow type within this module,
not much has to change!  Just the constructors; the pattern-match
engine caseNarrow; and apiNarrowFromNarrow, where we'd been taking
a shortcut using the fact that Narrow secretly was ApiNarrow.

Well, those and one call site to `specialNarrow`, where we newly
enforce an invariant we'd long assumed about what kind of `is:foo`
narrows we support.