unread: Use IDs, not emails, in getUnreadCountForNarrow.
As a bonus this eliminates one of the places using NULL_USER.

The case where that could have affected behavior was... if looking at
a PM conversation where one of the emails didn't correspond to any
user we knew, perhaps because the user had changed their email.  With
this fix, we'll correctly show an UnreadNotice "N unread messages"
banner if applicable.  With the old code, this selector would return 0
and so we wouldn't show the banner.