narrow: In isNarrowValid use only user IDs, not emails, for PMs.
This check is itself the last remaining place where we consume
the emails from a PM narrow and then try to find the users in our
data structures; everything else relies on the user IDs.  So there
are no longer any potential crashes that this check could prevent.

We're about to remove those emails entirely.  As part of that,
cut this check.