outbox [nfc]: Use IDs, not emails, to compute display_recipient.
This could conceivably affect behavior if we reached the point of
having composed, and attempting to send, a PM where we don't actually
have data for one of the recipients by email, but do by user ID,
or vice versa.

That shouldn't be possible, though, because the UI to show the
conversation in the first place would have already errored out by
that point.