api types: Use UserId in most of the API.
Specifically, use it in all the places that are purely an
"output position" (or nearly so), i.e. where we consume a value
from this type but don't try to put a value back into it.

This leaves out a few parts of the API we'll get to in
subsequent commits:

 * a few places which are "input positions", meaning the API code
   accepts values of this type instead of providing them -- in
   particular when we're providing user IDs as parameters;

 * a few places where we reuse the type for new values we construct
   for our own internal data structures, so that the type is in an
   input position there as well as an output position.

Those will require updating more of the app too, so that we can
supply UserId values.