compose: Give intended placeholder for all 1:1 threads.
I'm not sure it's actually possible to get a compose box for sending
to a deactivated user.  If it is, then previously we'd fall back to
the generic "Type a message" placeholder, and now we show the same
"Message ${name}" as we do for anyone else.

More importantly this gives us one fewer distinct thing to reason
about, and one fewer dependency on a duplicative large data structure.