user [nfc]: Eliminate bogus-data getSelfUserDetail; use getOwnUser.
These all appear inside the app's normal main UI, where we already
assume we have our normal app data; so we can just use `getOwnUser`,
which throws if it can't find the data it's supposed to return.

Initially I thought that `OwnAvatar` might need a workaround similar
to the one added in c1acd3a84 to `UserAvatar`, to deal with the same
bug where the UI seems to remain mounted and get updates after the
user has begun switching to another account and the data has changed
out from under it.  But empirically, things work just fine with this
straightforward version.

I think the point is probably that that bug specifically involves
changes to `state.accounts`, and is triggered by the use of selectors
like `getAuth` that rely on `state.accounts`.  This code doesn't, so
isn't affected.