autocomplete [nfc]: Stop using NULL_USER; be explicit about what we need.
Specifically we just need, for each option to present for
autocomplete, a user ID, an email, and a name.  That means that
when making up such an entry that isn't a real user, we can invent
just those fields, and don't need a generic "fake user" object to
supply dubious values for a bunch of other fields.

This still isn't a *great* way to handle these autocomplete
suggestions that aren't users; instead we should have some actually
appropriate bit of UI design, and data here that explicitly calls
for that.  But making it explicit like this what's going on here
will help make it easier to go about that change, too.