user [nfc]: Start switching from ownUser to ownUserId, in recipient.
In a lot of places where we pass around ownUser, we really just want
the user ID, in particular to recognize when some other user ID is or
isn't equal to it.  Passing just the user ID simplifies things because
it never changes, eliminating the possibility of being out of date.

Many of the places we pass the full ownUser, in fact, formerly had
just ownEmail, and we converted to ownUser in order to straddle code
that now used the user ID and code that still relied on emails.  So
converting further to just ownUserId is a way of completing the
email-to-user-ID conversion for that code.