avatar: Remove remaining URL construction for initial load.
This follows up on #4344 to address part of the remaining performance
regression from #4230.

We still construct URL objects in the `get` methods, lazily and
memoized; but never up front when simply taking in all the initial
data from the server, or the rehydrated data we've stored.  This
should save a further slice of time in the initial load.