store: Opt out of several `jsan` features we don't need.
This will probably give us some performance boost in our
replace/revive logic.

A tempting way to make a change like this would be to pass `false`
instead of this whole object -- but, examining the implementation,
that would effectively opt us back into the `refs` option, which
remotedev-serialize specifically wanted to set to false.