store: Use JSON instead of `jsan` for replace/revive logic.
The tricky part here is that, while we weren't relying on any of the
advertised extra features of `jsan`, we were relying on it to
recurse through our data and apply our `replacer` function [1].

JSON.stringify invariably calls the value's `.toJSON` method before
the supplied `replacer` function gets to have any say about how the
value should be serialized [2]. But Greg found a way to access the
value before the `.toJSON` transformation, and it's quite simple to
do with `JSON.stringify`'s API: we use `this`, which is the "holder"
object, and access the `key` on it.

The two things we use that have `toJSON` methods are `Immutable.Map`
(which makes an object-as-map of the data) and `URL` (which gives
the string representation). So, use `this[key]` to grab the
pre-`toJSON` values for those when we do our `instanceof` checks for
those.

All the other classes we check with `instanceof` clearly don't have
a `toJSON` method; we know this because we wrote them.

Also, add a test where `__serializedType__` is a key in an
Immutable.Map.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.60remotedev-serialize.60.20and.20.60jsan.60/near/1082842

[2] https://tc39.es/ecma262/#sec-serializejsonproperty (See steps 2
    and 3.)