avatar [nfc]: Drop the unused "shape" prop.
This isn't used, and makes the code more complex.  We always
effectively use the default of "rounded", never the alternative of
"square"; so just hard-wire the "rounded" behavior.

Back in 805d35d6f, fixing #3398, we took out the old "circle" option
here, which produced quite bad results when used.  The "square" option
could be OK in the right design context; but we don't actually use it,
and haven't since 57f45f439 a couple of years ago.  If we want that
functionality in the future, we can add it back in the relevant places
then.