msglist: Fix regression on tapping topic narrows in stream-narrow msglist.
In af1713a76, from #4339, a subtle bug was introduced by our use of
the `\x00` character in making keys for narrows (`keyFromNarrow`).
That character doesn't round-trip through being the value of the
`data-narrow` attribute in our message-list HTML, so we got an error
on pressing a topic-narrow header in the message list for a stream
narrow.

So, make it so that that attribute never contains `\x00` (or indeed
any other non-handled character [1]) by encoding it with our
`base64UtfEncode` function.

See discussion for how we arrived at this solution [2]. In
particular, it has the nice quality of also reducing our exposure to
bugs stemming from stream names (or topics, or users' email
addresses -- anything that ends up in the result of `keyFromNarrow`)
containing characters that aren't allowed in HTML-attribute values.

[1] https://html.spec.whatwg.org/multipage/syntax.html#attributes-2
[2] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M4339.20regression/near/1085234