replaceRevive tests: Test `stringify` and `parse` with our own data types.
In a recent commit, we stopped using remotedev-serialize's
`Serialize.immutable` factory to generate `stringify` and `parse`
based on a passed `customReplacer` and `customReviver`, instead
hard-coding `stringify` and `parse` in our own code using our custom
replacer and reviver logic.

Since we don't plan to use `Serialize.immutable` again, we don't
need tests to confirm that it does the right think with a provided
`customReplacer` and `customReplacer` -- so, delete those.

But we should ensure that our custom data types are tested just as
thorougly as the `Immutable.*` ones. So, add those to `data`.