user: Use user IDs more, and emails less, in UserList.
Continuing the conversion to UserItemById; and cutting a
reference to email along the way.

It'd be nice to go on to propagate this through the helper
`groupUsersByStatus`, but that can't trivially be done: it uses
the presence data, which is indexed by email.  We'll naturally
revisit this when we take care of migrating presence to user IDs.