replaceRevive [nfc]: Use function declarations.
Greg pointed out (in response to d75064571) [1] that the
`const foo = function foo() { ... }` form is repetitive, which it
is.

I'd given just a bit of reasoning for the repetitive form in
a0dcf4fdf, which I also gave in response to Greg's comment -- but
after more discussion, I was convinced that plain function
declarations (at toplevel, where definitions belong, and nontrivial
logic doesn't) are at least as good and possibly better. We'll fix
a0dcf4fdf in the next commit.

[1] https://github.com/zulip/zulip-mobile/pull/4367#discussion_r550311033