recent-pms: Reimplement getRecentConversations using the nice new data.
This is only a little shorter in line count; but I think it has
a good bit less complexity to take in.

It's also doing quite a lot less work: instead of iterating through
all the PMs (messages!) we know about, it iterates through just one
item per PM *conversation*, in data that we've efficiently
maintained as the messages came in.  That's the big payoff of this
server API feature.

Or, relatedly: because this is so much more efficient, we're able
to provide much more thorough data.  The old way was fed primarily
by `fetchPrivateMessages`, where we fetch the latest 100 PMs --
which for someone who uses PMs a lot could leave out conversations
as recent as yesterday, and include just a handful of conversations.
With current servers, the new way gets initialized from the user's
last 1000 PMs.

As a result, if you look at the PM-conversations screen with the
app before this change and then after it, you'll probably see a
lot more conversations.  For me on czo, I see about 4x as many,
extending far back into the past.

Fixes: #3133