collections: Add `mapOrNull`.
This replaces a pattern I think we've written out explicitly a fair
number of times, and which can be cleanly encapsulated instead.

For now this is the only function in this new file, but I think we
have a few other patterns with this flavor, which might be good to
add functions for here too.