unread tests: Convert stream-unreads reducer tests to whole-model style.
Instead of inspecting the internals of the state, the tests now look
at the state in a way that the app code in general, via selectors,
might look at it.  In particular, no selector cares (or should care)
about the order in which streams and topics appear.

Similarly, instead of constructing starting states by writing out the
internal representation, these tests now do so the same way the app
in general does: by building states up through the reducer.

I say "whole-model style" because I see this as a step toward
organizing our state-managing code in terms of a "model" that combines
the role of a reducer and selectors, and in which the representation
of the data is an internal implementation detail that the rest of the
app doesn't need to care about and can't even see.  See also previous
discussion in 7c697fb74.

In particular, I'd like to change the data structures we use for
tracking unreads, and moving that to be encapsulated inside a model
will make it much cleaner to change it.

As a useful bonus, the tests sure are a lot shorter this way!
I think they're easier to read and understand.