ZulipAppContainer [nfc]: Add thin `ZulipAppContainer` wrapper component.
And, within it, encapsulate the setting of the special `ref`
attribute to `NavigationService.appContainerRef`.

That interface change is made for convenience. I found that using
`React.forwardRef` to expose that capability to the caller would
have made it harder to do something unrelated, but that we'd like to
do soon.

(In particular, to call `createAppContainer` once, and in the right
time and place. We consider the appropriate pattern to be calling it
in `ZulipAppContainer`'s constructor and storing the result as an
instance property. But `React.forwardRef` takes a render function,
and a render function can't adequately describe `ZulipAppContainer`
while `ZulipAppContainer` depends on using an instance property.
"But wait," you might say, "that's not true! What about Hooks-based
components? After all, a Hooks-based component itself is basically a
glorified render function!" True, but there's a catch. The drop-in
equivalent for instance properties in Hooks-based components,
according to React [1], is the `useRef` hook. While they've
probabaly chosen the right name for that, I think it would just be
too confusing for us to use `useRef` alongside something we've used
just as rarely, with the very similar name `forwardRef`.)

[1] https://reactjs.org/docs/hooks-faq.html#is-there-something-like-instance-variables