ZulipMobile: Set initial route declaratively; remove "initial navigation".
In a recent commit, we ensured that none of React Navigation's
components get mounted until the store has rehydrated. The data from
rehydration is just what we need to tell React Navigation what the
"initial route" is. With that data, we can remove the awkward
`InitialNavigationDispatcher` logic, which would be difficult to
coordinate with the `NavigationService`'s initialization schedule in
React Navigation 5 [1], which we hope to upgrade to soon.

The mechanics of initializing the root navigator will be simpler
with React Navigation v5's component-based API. The initial route
name and params will just be passed as props to the appropriate
components; there isn't the kind of static configuration necessary
in React Navigation 4 via `createStackNavigator`.

[1] https://reactnavigation.org/docs/navigating-without-navigation-prop/#handling-initialization