PresenceStatusIndicator [nfc]: Use flex to center indicator in background.
It's easier to tell the thing to just center itself in its parent
than it is to give it an absolute positioning with an exact number
of pixels -- particularly if we ever decide we want to resize its
parent.

This also lets us remove the somewhat ambiguously named
`styles.status`.