topics [nfc]: Cut a few redundant type annotations.
Flow infers these just fine.

They don't add much for the reader, as the selectors that feed
these parameters are right there in the lines above and are
generally pretty clear.  (The reader might be surprised to learn
that `getMute` returns a list of pairs... but this annotation, as
`MuteState`, wasn't helping with that anyway.)

And by repeating the expected types for these selectors to return,
they add friction for changing them.  So take them out.

(A lot of our other createSelector calls could usefully get the
same treatment.)