user types: Fix Flow's typing of UserItemById component.
There seems to be a Flow bug where we weren't getting effective
type-checking of calls to this component -- a callsite could pass
arbitrary props, and fail to pass required ones, and Flow wouldn't
complain.  See discussion:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Flow.20and.20.60connect.60/near/1098203

We can fix that by explicitly providing the props type, rather than
using our `connect` with its subtle generic types that effectively
ask Flow to infer it in a particular way.

Moreover, while we're here, convert this `connect` call into a
Hooks-based component, using `useSelector`.  We still need the
wrapper component, unlike the typical `useSelector` idiom, because
we still have callsites invoking the inner component directly.
But even so, given all the type annotation we had to provide to
`connect` in this case, the `useSelector` version is a bit simpler.

See also #3768, where we ran into a very similar set of symptoms
which might involve the same underlying Flow bug.

Reported-by: Chris Bobbe <cbobbe@zulip.com>