user types: Use UserId in initial PM-conversations data.
This change in the API types is tied together with a small change in
our code that consumes it, for a bit of a silly reason: we want to
sort this array, and we'd rather do so in place rather than making an
unnecessary copy, and that means it can't be `$ReadOnlyArray`.

In principle this could be avoided by Flow having something
intermediate between plain `Array` and `$ReadOnlyArray`, with a name
like `$CovariantArray`.  Like `$ReadOnlyArray`, this would exclude
methods like `push` that accept a new item, and that therefore aren't
safe to use on an array that some other code might expect to have a
more specific type of element.  But unlike `$ReadOnlyArray`, it would
include `sort`, and `pop`, and other methods that mutate the array but
only provide elements as output, never taking them as input, and
therefore are just fine for that situation.

But Flow doesn't currently have that feature, so just treat this type
covariantly and update both the provider and consumers in lockstep.