UserItem [nfc]: Generalize type to accept fragments of a User.
This expresses just the properties that this component actually
needs from these objects.

A good thing, too, because in the case of showing autocomplete
suggestions, these are the only properties it's getting that are
at all explicitly chosen to be anything particularly reasonable;
the rest are bogus data from the NULL_USER value.

For that matter, even out of these three properties we use, one of
them is also bogus data from NULL_USER in that situation: namely,
`user_id`.  That's why we have to explicitly handle its possible
bogosity in `UserAvatarWithPresenceById`, the one place we use it.