types: Relocate some `$FlowFixMe`s and correct their TODOs.
As noted in the code comments, one mistake we were making (not seen
until now) was passing the wrong type parameter to `React$Ref`; we
should be passing a `React$ElementRef` type [1].

When we do that, the existing `$FlowFixMe`s (the ones with comments
about RN v0.63) are marked as unused, as we'd expect. But
unfortunately, at that point, `.current` is typed as
`any(implicit)`, which we definitely don't want. And that doesn't
seem to get fixed by upgrading to RN v0.63, so, remove the comments
about that new version. Greg suggests this may be caused by bugs in
Flow's special support for React [2].

So, mention our new knowledge that we need to be using
`React$ElementRef`, but retain some fixmes and comments alerting us
to the fact that `.current` is untyped and that we need to be
careful about using it.

[1] https://github.com/zulip/zulip-mobile/pull/4278#discussion_r512295663
[2] https://github.com/zulip/zulip-mobile/pull/4420#discussion_r561355599