unread [nfc]: Inline combineReducers, all in one step.
We've already done, for our overall reducer in src/boot/reducers.js,
the work of working through all the details of what combineReducers
does and unpacking it for an explicit set of sub-reducers.

So rather than repeat that, go in one step by basically copying the
structure of that resulting code.  Because we don't do the
maybeLogSlowReducer thing within this sub-reducer, we can simplify a
bit more by inlining
  applyReducer(key, subReducer, subState, action)
to just
  subReducer(subState, action)
.