redux [nfc]: Simplify combined reducer (6/n): unroll the main loop.
The list of our reducers is already right here in the same source
file, so we don't really need all this metaprogramming on it.  By
unrolling the loop and unpacking what it does, we make everything
more explicit, and in particular make it explicit to the type-checker
so that it's able to check the types and we no longer need any fixme
comments to turn it off.