generics [nfc]: Test BoundedDiff; fix some details in its jsdoc
Turns out the resulting type is actually exact, not inexact as the
jsdoc had said!  That's behavior inherited from $Diff.

Not sure if that changed at some point since I originally wrote this,
or if it was always this way -- could certainly be the latter, as
after all this didn't have tests. ;-)

Also don't try to characterize the behavior if the arguments aren't
read-only.  The read-only case is the only one we care about, and we
don't try to write tests for beyond that.