nav types: Accommodate an optional `route` prop in RouteParamsOf.
This lets us smoothly handle the case of LoadingScreen.

This solution feels a bit ad hoc, like there should be a more
general solution which I'm not seeing right now.  (Like instead of
dealing specially with void and null, as this does, the point should
be something like: the component accepts a broader type for the prop,
the navigator sees a more specific type as what it has to provide,
and that's all fine as long as the type the navigator undertakes to
provide is a subtype of the one the component expects to receive.)

But this should be perfectly correct in the cases where it does do
anything.  And I'm not sure there's a real use case for the more
general solution, anyway, so it's not worth spending too much time
trying to construct it unless we do find we need it later.