Lightbox: Redo slide animation of header and footer.
The previous way we did this animation depended on Lightbox knowing
(or guessing! NAVBAR_SIZE was a wrong guess) the heights of
LightboxHeader and LightboxFooter [1]. This wasn't ideal, partly
because it meant coordinating between `Lightbox` and those
components, but also because the most reasonable, bug-free way to
know the height of something is to set an explicit height, and that
would mean limiting our options for being faithful to the user's
system font-size setting. I also want to bring the footer into a
more consistent appearance with the header, and as part of that I
want to remove its explicit height.

The `LayoutAnimation` API [2] is described in the "Animations"
doc [3] as being powerful but not as granular as the plain Animated
API [4]; it lets you animate "layout updates without bothering to
measure or calculate specific properties" -- which is just what
we're looking for.

The new animation is meant to roughly match the old; the header and
footer still slide in and out, though it's probable that the
position vs. time graph looks slightly different now than before.

See also discussion [5].

[1] I tried using values like "0%" and "100%" for `from` and `to`
    (`from` and `to` being the values that get plugged into
    `transform.translateY`), which would mean we wouldn't have to
    know the heights. But using percentages isn't supported yet; see
    https://github.com/facebook/react-native/issues/13107.

[2] https://reactnative.dev/docs/layoutanimation

[3] https://reactnative.dev/docs/animations#layoutanimation-api

[4] https://reactnative.dev/docs/animated

[5] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.60View.60's.20.60onLayout.60.20prop/near/1110130