lightbox header/footer: Make buttons more consistent.
The "x" button in the header was putting a brand-colored rectangle
around the icon when it was pressed, which seems jarring and
artificial. We often want this rectangle, in cases where its shape
follows contours that are visible when the button is not touched
(e.g., stream and topic items in the unreads list have their own
clear rectangles that just become highlighted when you press them).
But here, it's not following any contours like that; the rectangle
has seemed kind of random and unnecessarily sharp.

The three-dots button was giving touch feedback in a much nicer way:
slightly dimming just the icon itself, and no rectangle around it.
So, make that happen for both buttons. The dimming effect was small
enough to be almost unnoticeable, so the new touch-feedback response
(turning it from "white" to "gray") increases it a bit.

Neither button was giving the touch target the minimum 48 x 48 px
size recommended by Material Design [1]; so, do that with the
`hitSlop` prop [2]. Passing one number like this adds that amount of
additional distance outside the element in which a press can be
detected. So, with a size of 28, the new width and height of the
pressable area are 10 + 28 + 10 = 48.

This is our first experimentation with the Pressable API [3], just
added in React Native v63.

[1] https://material.io/design/usability/accessibility.html#layout-and-typography
[2] https://reactnative.dev/docs/pressable#hitslop; also available
    on `View` and other components.
[3] https://reactnative.dev/docs/pressable