lightbox header/footer: Stop constricting the buttons' touchable area.
The extension of a view's touchable area caused by the `hitSlop`
prop can be restricted by a few factors [1] [2]:

"""
The touch area never extends past the parent view bounds and the
Z-index of sibling views always takes precedence if a touch hits two
overlapping views.
"""

We don't want the buttons' parent view bounds to reduce the
touchable area at all. So, to be sure they don't, grow the parent
view bounds just enough.

Horizontally, we're already in good shape and we don't need to do
anything; the parent has a horizontal padding of 16, which keeps the
touchable area away from the right edge of the screen.

Vertically, the parent has 8 padding above and 8 padding below.
Increase these to 12 (the value we've chosen for `hitSlop`) by
giving the button an upper margin of 4 and a lower margin of 4.

See discussion [3].

[1] https://reactnative.dev/docs/pressable#hitslop
[2] https://reactnative.dev/docs/view#hitslop
[3] https://github.com/zulip/zulip-mobile/pull/4442#discussion_r567137485