nav: Navigate to 'main-tabs' instead of 'loading'.
We've just made it safe to do this by having `MainTabsScreen` show
the full-screen loading indicator if it doesn't have server data.

Now, the time between LOGIN_SUCCESS/ACCOUNT_SWITCH and seeing the
server data on MainTabsScreen is spent on MainTabsScreen itself, but
in the branch that shows the loading indicator. Before, it was spent
on LoadingScreen, which meant some confusing [1] logic about
navigating to and from the 'loading' route.

The 'loading' route ceases to be something we'd ever want or need to
navigate to, and we'll remove it in an upcoming commit.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M4453.20.22Tried.20to.20use.20NavigationService.20before.20appContaine.2E.2E.2E/near/1111808