MainStackScreen [nfc]: Rename, from `AppNavigator`.
Soon, we'd like to nest this whole thing as a screen under a new
stack navigator, and set that stack navigator to use `mode="modal"`.
This is according to React Navigation's guide to implementing
full-screen modals [1].

The example in that guide uses the name `MainStackScreen` for this
kind of thing; so, might as well follow that.

Rename `AppNavigatorParamList` to `MainStackNavigatorParamList`, not
`MainStackScreenParamList`, to emphasize the component's role as a
navigator, not as a screen -- because the "param list" is something
that applies to the screens on the navigator. (Like we did with
`MainTabsNavigatorParamList`, etc., seen together in
src/nav/globalTypes.js.)

[1] https://reactnavigation.org/docs/modal/