ZulipAsyncStorage: Implement `removeItem`.
redux-persist expects the `storage` object from its config to have a
`removeItem` method; specifically,
src/third/redux-persist/purgeStoredState.js uses it.

I guess we've never exercised `purgeStoredState` or any other code
that depends on `removeItem` being implemented on
`ZulipAsyncStorage` -- but we're about to, and we'll want it to work
smoothly.

The homemade libdef was missing `removeItem`, so, add it there.

See discussion at
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M4458.3A.20.22t.2Eget.20is.20not.20a.20function.22.20on.20state.2Enarrows.20at.20sta.2E.2E.2E/near/1119801.