MentionUserNotSubscribed: Add entry and exit animation to the component.
Use a nice animation for this component.

Use LayoutAnimation API to animate the component.
Make use of componentDidMount and componentWillMount
to make the component animate in both mount and
unmount state.

Make use of absolute positioning to make sure that the
warning message does not move the Webview up and down.
Also, now the warning message overlay on existing message in
the z-direction.

Discussion of this in CZO [1]

[1] https://chat.zulip.org/#narrow/stream/48-mobile/topic/UI.20tweaks

Fixes part of #4224