AuthScreen: Get `props.realm` from `serverSettings`, not Redux.
The auth screen is meant to log someone into a specific realm. The
identity of that realm is pretty clear, or it should be: it's the
realm that was successfully validated, and connected to, in the
`RealmInputScreen`. If the auth screen takes any other realm as the
realm to use, that's a bug. It'd be wrong if the realm changed
halfway through using the auth screen, and subscribing to a value
from Redux is just inviting that to happen.

We could simply have `AuthScreen` take the client-validated realm
URL as a route parameter. But, better yet: we use
`serverSettings.realm_uri`, which is explicitly the realm URL that
the server identifies for itself.