StoreProvider: Set server version tags on global scope, not one-off scope.
We discovered [1] that we'd only been including the server version
tags when making a `logging.foo` call. This excludes the large
volume of Sentry events that come from uncaught exceptions.

So, set the tags on the global scope, and be sure to keep them
up-to-date by listening for changes to the server version stored in
Redux.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M4458.3A.20.22t.2Eget.20is.20not.20a.20function.22.20on.20state.2Enarrows.20at.20sta.2E.2E.2E/near/1119488
[2] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Custom.20tags.20for.20Sentry's.20exception.20tracking/near/1121623