modelTypes: Move `recipient_id` higher up in `Message`.
Above the line that introduces "properties that behave differently
for stream vs. private messages".

As the comment says, we don't use this property. But if we did, I'm
not sure we'd want to describe it as behaving differently for stream
messages and PMs.

Here's the current doc for it [1]:

"""
`recipient_id`: integer

A unique ID for the set of users receiving the message (either a
stream or group of users). Useful primarily for hashing.
"""

While it's not super clear, it seems like any difference in its
behavior between stream and PM messages looks like an implementation
detail in the service of being "useful primarily for hashing".

If we find that we need to use this, and we also run into problems
like unwanted collisions in this field between PM and stream
messages...then I suppose we could go back to treating this as
different between PM and stream messages. But I imagine the way we'd
do that is to put `recipient_id: number` on both `PmMessage` and
`StreamMessages` (these types will come later in this series). Which
after all wouldn't be helpful for type-checking.

[1] https://zulip.com/api/get-messages