flow: Upgrade to v0.125.
Normally, we wouldn't upgrade Flow except in the same commit as we
upgrade React Native. That's because the code for a React Native
release (the code in node_modules/react-native) is only guaranteed
to pass type-checking with the Flow version that was used to check
the code at that release.

Given that this commit breaks from our normal practice, it's
unsurprising to see a Flow error in a file in
node_modules/react-native. Suppress type-checking in that specific
file (it's just one file), noting that it's not a file that we use
directly, nor does it seem likely that the suppression propagates to
a type-checking lack in our own project. (The most relevant-looking
site I've seen, our use of `UIManager` in src/ZulipMobile.js, still
has its types after this change.)

The nice thing we get in Flow v0.125 is that `$FlowIssue` and
`$FlowExpectedError` are added as default suppression comments [1],
which will allow us to remove the `suppress_comment` lines in our
Flow config that define those. And that's desirable because
`suppress_comment` has been removed from Flow in v0.127 [2], and
we'll have to take that removal along with the Flow version we take
with RN v0.64.

Best, then, to handle as much of this change as we can before the
main RN v0.64 upgrade commit.

In our own code, we have to slightly relocate a suppression comment
for facebook/flow#8276; so, do that.

[1] https://github.com/facebook/flow/blob/master/Changelog.md#01250
[2] https://flow.org/en/docs/config/options/#toc-suppress-comment-regex