ZulipButton: Extend button touch target to 48px.
Made the use of hitSlop prop and passed it from
MentionedUserNotSubscribed to ZulipButton.

Touch target now extends to 48px by specifying
top and bottom properties of hitSlop prop to
6px each. Since the height of the button is 32,
so adding the hitSlop to 6 at both top and bottom
sums it up to 48.
Link to the doc - https://support.google.com/accessibility/android/answer/7101858?hl=en

Discussion of this in CZO [1]

[1] https://chat.zulip.org/#narrow/stream/48-mobile/topic/UI.20tweaks

Fixes part of #4224