flowconfig [nfc]: Follow RN in removing `deprecated-utility` lint.
Done now, to follow the template-app change in
facebook/react-native@fcffb961f, on the path to the RN v0.64
upgrade.

That commit reports that this lint rule is no longer needed because
it's turned on by default. This is corroborated by the Flow
changelog for v0.98.0 [1], and I verified it experimentally -- using
`$Supertype` (one of the deprecated utility types this rule is meant
to catch [2]) is flagged as an error with or without this line in
our config.

[1] https://github.com/facebook/flow/blob/master/Changelog.md#0980
[2] https://flow.org/en/docs/linting/rule-reference/#toc-deprecated-utility