accountsReducer tests: Clarify expectations and data flow in LOGIN_SUCCESS.
In the spirit of adc4c8fb6, describe first the input and then the
expected result.

The new "without clobbering anything else" logic isn't thoroughly
tested because we don't test a variety of different inputs. In
theory, the reducer could be clobbering the value at `.foo` with
something, but the test doesn't pick it up because the new value
happens to be exactly the same as the `.foo` we chose for the input.
In practice, the reducer is clobbering `ackedPushToken` with `null`,
and the test doesn't notice because we've chosen an input that
happens to have `null` at `ackedPushToken`. We'll fix that soon.