ActionSheets: Split showActionSheet into multiple functions
This splits showActionSheet into a showHeaderActionSheet and a
showMessageActionSheet function, since the two of them conceptually
require different arguments and do different things.

Right now, they take the same set of arguments, but this commit sets us
up to change that in the future, and tighten up the typing more.