ActionSheets: Remove ButtonCode, pass around objects instead
This removes the ButtonCode infrastructure, and instead has the code for
constructing button lists pass around ButtonDescription objects.

This will enable us to split `ButtonDescription` into a `HeaderButton`
and `MessageButton` type (or a `Button<Message | Header>` type, more
likely), and have each of those take only the arguments it needs - for
instance, we can avoid passing `startEditMessage` callback to a
`Button<Header>`, since we know that'll never be used.

This also lets us avoid keeping around a global list of all possible
buttons, while still being able to share button code across multiple
actionsheets in a typesafe way (for instance, the cancel button).

The main downside is that our tests now rely on the `title` of the
button to check if the correct buttons have been inserted, since we no
longer have a `ButtonCode` to compare. At first glance, this seems bad,
but I think it's actually OK:

* We apply i18n in the rendering step, so we don't need to worry about
  the strings being translated - there are no calls to `_` in the
  codepath that the tests are checking.
* We're not really much worse off than we were before - previously,
  changing the name of a button in the code would necessitate changing
  the tests, now changing the title will require that. While a title
  change is probably more common, I don't think we're at any increased
  risk of "stringly typed" typo bugs - if we change a title, the tests
  will start failing and we'll need to go through and patch it up, but
  none of the non-test logic relies on the button titles.