MainTabsScreen [nfc]: Use `useSelector` for `haveServerData`.
In our understanding of the reason to use `connect` (see the long
code comment above the `connect` call), it shouldn't be necessary to
use `connect` to grab any particular value; the essential thing is
that the component be wrapped with a `connect` call.

So, keep doing that, but use `useSelector` to get `haveServerData`.