MainTabsScreen [nfc]: Use `compose`.
I always forget the order in which the functions are applied, but I
think this is correct. From the doc [1]:

"""
(arguments): The functions to compose. Each function is expected to
accept a single parameter. Its return value will be provided as an
argument to the function standing to the left, and so on. The
exception is the right-most argument which can accept multiple
parameters, as it will provide the signature for the resulting
composed function.
"""

[1] https://redux.js.org/api/compose