ios: Update "Brand" colorset with new brand color!
This is the RGBA representation of `hsl(222, 99%, 69%)`, which we've
set as the BRAND_COLOR value in the JavaScript layer. I would put
this in a code comment, but code comments aren't supported in JSON,
which is unfortunate.

We introduced this iOS colorset in 86843a122, to help us get to a
world where it's easier to have the native iOS color definitions
match those in the JavaScript layer.

We haven't yet completed that work by using RN's new PlatformColor
API (which is available to us now because we're using RN v0.63). But
anyway, now that we've started using the new brand color in the
JavaScript layer, we should change this one to match.

Checking #4279 again, we didn't add a counterpart native Android
file that defines the brand color, and I didn't find an existing
one, so I expect we don't need to make this kind of change for
Android.  (The color does appear in the native Android code
directly, in a call to a `setColor` method for giving our
notification icon a color.)