Setup VideoPlayer
Following the same implementation of Lightbox and using LightboxHeader.

The `Video` component accept several props, most important is the
`source`, which take the video uri and the headers -if exists-.

Wrapping the `Video` with `Touchable` to handle show/hide header,and
because unlike `PhotoView` the `Video` doesn't have `onTap` props.

Handle video error by show a message on the screen