UploadedAvatarURL [nfc]: Remove `if (match !== null)` check.
And have the `.replace` call take responsibility for executing the
pattern.

`.replace` already has the behavior of doing nothing if a match
isn't found, and we now pass an equivalent pattern to `.replace`. So
this should be NFC as long as we've correctly written the replacer
function to match the previous replacement logic, and I believe we
have.