SwitchRow [nfc]: Rename, from `OptionRow`.
The new name is more specific about the component's job: to show a
labeled switch.

The old name might also have suggested that the component represents
one item among a set of closely related items, which it doesn't,
necessarily [1]. It would have done in cases where we'd put it to
use for showing a radio button [2], a checkbox [3], or a menu item
[4], but those aren't possible in its current implementation.

And that's probably for the best; this rename narrows the
component's job, instead of widening it, on purpose. Radio buttons,
checkboxes, and menu items are the most likely-looking things we
might expand this component's job to cover -- but an item of those
types probably wants to assume it belongs to a set of closely
related items of its type (a checkbox in a list of checkboxes, etc.)
when it chooses its implementation.

[1] https://material.io/components/switches: "Switches toggle the
    state of a single item on or off."

[2] https://material.io/components/radio-buttons: "...one option
    from a set".

[3] https://material.io/components/checkboxes: "...one or more items
    from a set".

[4] https://material.io/components/menus#usage: "...from multiple
    options".