SelectableOptionRow [nfc]: Widen interface to match name (2/5).
`name` was used for the translated name of the language, and it
appeared just under the native name. For example, if you're viewing
the app in English, you'd see an item like

--------
Français
French
--------

where the first line is `nativeName` and the second is `name`.

An interface like this seems useful for callers, except not all
callers will be dealing with locale names. So, take `name` and give
it the more generic name `subtitle`; we'll change `nativeName` to
`title` in the next commit.