Settings: Add "Open links with in-app browser" setting.
This will let people choose whether they want to use the in-app browser,
or their external browser app.

This defaults to enabled on Android, and disabled on iOS.

The setting for this could probably be improved - See the discussion [1]
on CZO for what this looks like in other apps. I've opted to use a
simple switch, since that's what we're already using in this UI, and I
don't want to redesign it right now. I expect we will want to redesign
it in the future, but given that we have so few settings right now, I'm
not too worried about the settings screen being confusing/overwhelming.

Another thing that I'd like to improve is the icon - I'm using the
Feather icon for Google Chrome, since the only other browser-related
icons I could find were the globe icon we're using for the language
setting in this same screen, and the compass icon, which is less
recognizable than the Chrome one. I'm unhappy with using a branded icon
here, but Feather doesn't give us very many options, sadly.

[1]: https://chat.zulip.org/#narrow/stream/48-mobile/topic/in-app.20browser/near/1168259